Moves Configuration class into configuration.py
The module can now be imported when necessary. In this way, we can build
new tools which share the configuration file and parsing with
teacup-storm.py.
